+++ ldap/client.go
@@ -215,7 +215,8 @@ type Attribute struct {
 // the WithGroups option is specified, it will also return the user's groups
 // from the directory.
 //
-// Supported options: WithUserAttributes, WithGroups, WithDialer, WithURLs
+// Supported options: WithUserAttributes, WithGroups, WithDialer, WithURLs,
+// WithLowerUserAttributeKeys, WithEmptyAnonymousGroupSearch
 func (c *Client) Authenticate(ctx context.Context, username, password string, opt ...Option) (*AuthResult, error) {
 	const op = "ldap.(Client).Authenticate"
 	if username == "" {
@@ -273,7 +274,11 @@ func (c *Client) Authenticate(ctx context.Context, username, password string, op
 			return nil, fmt.Errorf("%s: failed to get user attributes: %w", op, err)
 		}
 		for _, a := range attrs {
-			userAttrs[a.Name] = a.Vals
+			name := a.Name
+			if c.conf.LowerUserAttributeKeys || opts.withLowerUserAttributeKeys {
+				name = strings.ToLower(a.Name)
+			}
+			userAttrs[name] = a.Vals
 		}
 	}
 	if !opts.withGroups && !c.conf.IncludeUserGroups {
@@ -285,7 +290,14 @@ func (c *Client) Authenticate(ctx context.Context, username, password string, op
 	}
 
 	if c.conf.AnonymousGroupSearch {
-		if err := c.conn.UnauthenticatedBind(userDN); err != nil {
+		// Some LDAP servers will reject anonymous group searches if userDN is
+		// included in the query.
+		dn := userDN
+		if c.conf.AllowEmptyAnonymousGroupSearch || opts.withEmptyAnonymousGroupSearch {
+			dn = ""
+		}
+
+		if err := c.conn.UnauthenticatedBind(dn); err != nil {
 			return nil, fmt.Errorf("%s: group search anonymous bind failed: %w", op, err)
 		}
 	}
@@ -707,8 +719,13 @@ func (c *Client) getUserDN(bindDN, username string) (string, error) {
 	}
 	var userDN string
 	if c.conf.UPNDomain != "" {
-		// Find the distinguished name for the user if userPrincipalName used for login
-		filter := fmt.Sprintf("(userPrincipalName=%s@%s)", escapeValue(username), c.conf.UPNDomain)
+		// Find the distinguished name for the user if userPrincipalName used for login, or sAMAccountName if enabled.
+		var filter string
+		if c.conf.EnableSamaccountnameLogin {
+			filter = fmt.Sprintf("(|(userPrincipalName=%s@%s)(sAMAccountName=%s))", escapeValue(username), c.conf.UPNDomain, escapeValue(username))
+		} else {
+			filter = fmt.Sprintf("(userPrincipalName=%s@%s)", escapeValue(username), c.conf.UPNDomain)
+		}
 		result, err := c.conn.Search(&ldap.SearchRequest{
 			BaseDN:       c.conf.UserDN,
 			Scope:        ldap.ScopeWholeSubtree,
@@ -719,9 +736,10 @@ func (c *Client) getUserDN(bindDN, username string) (string, error) {
 		if err != nil {
 			return userDN, fmt.Errorf("%s: LDAP search failed for detecting user (baseDN: %q / filter: %q): %w", op, c.conf.UserDN, filter, err)
 		}
-		for _, e := range result.Entries {
-			userDN = e.DN
+		if len(result.Entries) != 1 {
+			return "", fmt.Errorf("%s: LDAP search for user 0 or not unique", op)
 		}
+		userDN = result.Entries[0].DN
 	} else {
 		userDN = bindDN
 	}